% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{assign_weights}
\alias{assign_weights}
\title{Assign Neural Networks Weights from Array}
\usage{
assign_weights(weights, weightdim, avec)
}
\arguments{
\item{weights}{neural network weights in list structure}

\item{weightdim}{dimensions of neural network weights}

\item{avec}{new weight array}

\item{model}{Model Object}
}
\description{
This is a function to assign neural network weights from a weight array
provided a neural network structure
}
