% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{autogen_cnn}
\alias{autogen_cnn}
\title{Automatically Generate Convolutional Neural Network for Coil Parameterization}
\usage{
autogen_cnn(dfs, n.s, params)
}
\arguments{
\item{dfs}{transformed data}

\item{n.s}{number of coil presentations}

\item{params}{number of coil values}
}
\description{
This is a function for automatically generating a convolutional neural network
for a transformed dataset in the unique manner required for coil parameterization
}
