% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coilfunctions.R
\name{buildcoil}
\alias{buildcoil}
\title{Build Complex Valued Coil}
\usage{
buildcoil(n.s, sym = FALSE)
}
\arguments{
\item{n.s}{number of presentations}

\item{sym}{is the coil symmetrical?}
}
\description{
This function generates a complex valued coil based on user specifications of size
and symmetry
}
