% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{eval_params}
\alias{eval_params}
\title{Evaluate Parameters by Populating and Running Coil Directly}
\usage{
eval_params(avec, inputlist, outputs, sel = seq(1, lookforward))
}
\arguments{
\item{avec}{parameter array values}

\item{inputlist}{input list to track where parameters should be assigned.}

\item{outputs}{objective values to evaluate against}

\item{sel}{which values to use in loss function evaluation}
}
\value{
total error
}
\description{
This is a function to pass parameter values into a coil and evaluate its
fitness for a number of inputs and outputs
}
