% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{gen_in_out}
\alias{gen_in_out}
\title{Easily generate input and output lists}
\usage{
gen_in_out(dfs, xsamps)
}
\arguments{
\item{dfs}{transformed data}

\item{xsamps}{sampled index}
}
\value{
input list and output matrix
}
\description{
This is a function to generate input and output lists based on some specified
sampling index
}
