% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{get_params}
\alias{get_params}
\title{Evaluate Neural Network to Get Coil Parameters}
\usage{
get_params(input)
}
\arguments{
\item{input}{input values for neural network}

\item{model}{Model Object}
}
\value{
coil parameter values

coil rotation values

coil start values
}
\description{
This is a function to evaluate the neural network for some inputs to get
coil parameters
}
