% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{initialize_swarm}
\alias{initialize_swarm}
\title{Initialize Particle Swarm}
\usage{
initialize_swarm(swarm_size, L = length(avec), locfac = 0.6, type = "neural")
}
\arguments{
\item{swarm_size}{number of particles in swarm}

\item{L}{number of dimensions needed for swarm}

\item{type}{what type of optimization is the swarm used for?
\itemize{
\item{"neural"}{ Optimize weights of convolutional neural network to predict rotation, start values, and coil parameterizations from inputs}
\item{"direct"}{ Optimize unique rotations and start values for multiple outputs using the same coil parameterization}
}}

\item{loc_fac}{locality factor}
}
\description{
This is a function to initialize a particle swarm either for direct optimization
or neural network optimization
}
