% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{initialize_swarm_full}
\alias{initialize_swarm_full}
\title{Initialize Particle Swarm for Full Coil Solution}
\usage{
initialize_swarm_full(swarm_size, L, locfac = 0.6, setrots = NULL)
}
\arguments{
\item{swarm_size}{number of particles in swarm}

\item{L}{number of dimensions needed for swarm}

\item{setrots}{use fixed rotation values?}

\item{loc_fac}{locality factor}

\item{type}{what type of optimization is the swarm used for?}
}
\description{
This is a function to initialize a particle swarm to determine coil rotations, starts, and parameters
}
