% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{invertscaling}
\alias{invertscaling}
\title{Invert Scaling}
\usage{
invertscaling(x, scaler, range)
}
\arguments{
\item{x}{scaled data}

\item{scaler}{values with which data was scaled}

\item{range}{desired scaling range}
}
\value{
unscaled data
}
\description{
This function inverts data scaling provided some scaled data, the values the
data was scaled with, and the specified range
}
