% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{layer_activation_seagull}
\alias{layer_activation_seagull}
\title{Seagull Activation}
\usage{
layer_activation_seagull(x, alpha, decay = 1000)
}
\arguments{
\item{x}{what to compose new layer instance with}

\item{alpha}{float >= 0. Negative slope coefficient.}

\item{decay}{smoothing value to prevent rapid approach of maximum}
}
\description{
This is a unique activation function for the parameterization of coil values
with a neural network
}
