% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coilfunctions.R
\name{llen}
\alias{llen}
\title{Find Effective Presentation Magnitude for Real-Valued Probability}
\usage{
llen(varlarr)
}
\arguments{
\item{valarr}{complex valued array}
}
\value{
real-valued lengths
}
\description{
This function translates complex-valued probability into real values
}
