% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{lookwindow}
\alias{lookwindow}
\title{Rearranges Data for Convolutional Time Series Training}
\usage{
lookwindow(df, lookback, lookforward, predictors, objective)
}
\arguments{
\item{df}{data frame with single feature row}

\item{lookback}{how many timesteps should we look back on}

\item{lookforward}{how many timesteps in the future do we what to predict}

\item{predictors}{which is the objective feature}
}
\value{
list of transformed data
}
\description{
This transforms a data frame into arrays of lookbacks and lookforwards
}
