% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{lossfun}
\alias{lossfun}
\title{Generic Mean Squared Loss Function}
\usage{
lossfun(actual, predicted)
}
\arguments{
\item{actual}{objective values}

\item{predicted}{model values}
}
\description{
This is a function used to evaluate loss based on actual vs predicted
}
