% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coilfunctions.R
\name{lrad}
\alias{lrad}
\title{Find Effective Presentation Magnitude for Prioritization}
\usage{
lrad(varlarr)
}
\arguments{
\item{valarr}{complex valued array}
}
\value{
real-valued lengths
}
\description{
This function finds the effective length of presentations for prioritization
during coil dynamics
}
