% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coilfunctions.R
\name{normalize}
\alias{normalize}
\title{Perform Normalization of Complex Values}
\usage{
normalize(comvec, GoalR, Goali)
}
\arguments{
\item{comvec}{complex valued array}

\item{GoalR}{real value of objective point along complex unit circle}

\item{GoalI}{imaginary value of objective point along complex unit circle}
}
\value{
normalized complex value array
}
\description{
This function normalizes an array of complex values to a specific point along
the complex unit circle
}
