% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{optf}
\alias{optf}
\title{Fitness Function for Full Coil Solution}
\usage{
optf(outs, out_goal, slseq)
}
\arguments{
\item{outs}{modeled values}

\item{out_goal}{objective values}

\item{slseq}{objective points selection}
}
\description{
This is a function to evaluate the particle swarm for a full coil solution
}
