% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{pop_coil}
\alias{pop_coil}
\title{Populate and Run Coil for Provided Patameters}
\usage{
pop_coil(input, readout = F)
}
\arguments{
\item{input}{input values for neural network}

\item{readout}{should the neural network results be displayed?}
}
\value{
coil values
}
\description{
This is a function to populate and run a coil by evaluating a neural network
for a set of inputs
}
