% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{pop_coil_full}
\alias{pop_coil_full}
\title{Populate and Run Coil for Rotations, Start Values, and Coil Parameters}
\usage{
pop_coil_full(avec)
}
\arguments{
\item{avec}{array of rotation, start values, and coil parameters}
}
\value{
coil result
}
\description{
This is a function to populate and run a coil provided the rotations, start values, and coil parameters
}
