% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coilfunctions.R
\name{runcoil}
\alias{runcoil}
\title{Simulate Coil Dynamics}
\usage{
runcoil(RandVec, rotvals, startvals, prival = NULL)
}
\arguments{
\item{RandVec}{complex valued array of coil parameters}

\item{rotvals}{three real values giving coil rotations}

\item{startvals}{complex valued array giving coil starting values}

\item{prival}{what prioritization function to use?}
}
\value{
real-valued probability dynamics over time

complex-valued coil dynamics
}
\description{
This functions runs the generated coil provided the parameterizations
}
