% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{scaledata}
\alias{scaledata}
\title{Scale Data}
\usage{
scaledata(x, range)
}
\arguments{
\item{x}{data array}

\item{range}{desired scaling range}
}
\value{
scaled data
}
\description{
This function scaled data to a specified range based on minimum and maximum
}
