% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{scalelist}
\alias{scalelist}
\title{Scale Data Frame}
\usage{
scalelist(dfw, predrange = c(0, 1), objrange = c(0.2, 0.6))
}
\arguments{
\item{dfw}{data frame with single feature row}

\item{predrange}{what range should the predictor values be scaled to?}

\item{objrange}{what range should the objective values be scaled to?}
}
\value{
scaled data frame

scaling values
}
\description{
This applies scaling to an entire data frame with the ability to scale to
different ranges for predictor vs objective features, and saves the scaling
values.
}
