% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{step_swarm}
\alias{step_swarm}
\title{Step through Particle Swarm}
\usage{
step_swarm(
  swarm_size,
  L = length(avec),
  w = 0.9,
  g_p = 0.4,
  g_g = 0.4,
  type = "neural"
)
}
\arguments{
\item{swarm_size}{number of particles in swarm}

\item{L}{number of dimensions needed for swarm}

\item{w}{momentum factor}

\item{g_p}{global pull factor}

\item{g_g}{local pull factor}

\item{type}{what type of optimization is the swarm used for?
\itemize{
\item{"neural"}{ Optimize weights of convolutional neural network to predict rotation, start values, and coil parameterizations from inputs}
\item{"direct"}{ Optimize unique rotations and start values for multiple outputs using the same coil parameterization}
}}

\item{loc_fac}{locality factor}
}
\description{
This is a function to step through a particle swarm either for direct optimization
or neural network optimization
}
