% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{step_swarm_full}
\alias{step_swarm_full}
\title{Step through Particle Swarm for Full Coil Solution}
\usage{
step_swarm_full(swarm_size, L, w = 0.9, g_p = 0.4, g_g = 0.4, setrots = NULL)
}
\arguments{
\item{swarm_size}{number of particles in swarm}

\item{L}{number of dimensions needed for swarm}

\item{w}{momentum factor}

\item{g_p}{global pull factor}

\item{g_g}{local pull factor}

\item{setrots}{use fixed rotation values?}
}
\description{
This is a function to step through a particle swarm to determine coil rotations,
starts, and parameters
}
