% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfunctions.R
\name{transform_to_params}
\alias{transform_to_params}
\title{Transform Array into Coil Parameters}
\usage{
transform_to_params(avec, inputlist)
}
\arguments{
\item{avec}{parameter array values}

\item{inputlist}{input list to track where parameters should be assigned.}
}
\value{
rotations and start values
}
\description{
This is a function to transform an array into coil parameters directly.
For use in direct particle swarm optimization, unique parameters are generated
for each input.
}
